Added an admin module for ldap based email settings (based on emailadmin) but reduced to those functions, which are needed to manage ldap entries, which rely only on standard schemas (core, qmail) which are not considered "experimental" and without the need to add a separate schema.
This modul will be extended whenever possible and approved for more values.

Main goal here is: Ability to easily use eGW Admin for account management with
LDAP beackend and to comply with those schemas, the distros are delivering.

The link for this module will be shown in addition to Lars' EmailAdmin if LDAP
is used as account storage. So it will not break existing installation ;-)


git-svn-id: http://svn.egroupware.org/egroupware/trunk@16621 923e3c2e-ae13-0410-ba18-ced7211e33de