Patch from Jan
It is the simple patch to prevents the segfaulting of horde on ical with
spaces in the begin:vevent etc lines. (just \W\w replaced by \S if I
remember well at one place..)

The second patch is for getting rid of some php5 warnings when it
encounters the array_merge() in setParameter(). (only a simple cast to
an array).


git-svn-id: http://svn.egroupware.org/egroupware/trunk@20536 923e3c2e-ae13-0410-ba18-ced7211e33de