Improved cache handling in egw_link class, to cope with excessive multiple reads of entries from the database: Applications can call
egw_link::set_cache($app,$id,$title,$file_access=null)

from their search or read method, to eliminate the need to query the
entries again, when the egw_link class, link widget or links stream wrapper
needs title or file_access values later.
This offloads the caching to the link class, and improves performance a
lot, specially for infolog.
The cache is stored in the session and modified or deleted items get
removed, when the link class get notified about that anyway.


git-svn-id: http://svn.egroupware.org/egroupware/trunk@26236 923e3c2e-ae13-0410-ba18-ced7211e33de