Tracker #2354. Add a parameter in the news_admin sitemgr module to select a news.
This allows to create content as a news item and to save the effort of
duplicating this content as a separate, distinct web page in the site tree. The
distinct web page can just use the news module and select the news item to be
displayed on this page.

As a separate szenario, this comes handy for news by time, to be published on the
main page of the website. Now, multiple events can be filed and made available as a
news item and the top-priority news (which ndoes no more need to be the last
recent news) can be displayed on the start page, being
automatically replaced by less important news, once the most important news is
timed out.


git-svn-id: http://svn.egroupware.org/egroupware/trunk@31520 923e3c2e-ae13-0410-ba18-ced7211e33de