Add support for filtering records to be exported.
Filterable fields are attempted to be autodetected by using the exportable fields.
Records can be filtered by fields with type select,select-cat,select-account,date,date-time (according to egw_record class) 
only at this time.  Filters are saved in the definition and used with scheduled exports.  They are also available to the
user for modification in the export dialog.



git-svn-id: http://svn.egroupware.org/egroupware/trunk@41134 923e3c2e-ae13-0410-ba18-ced7211e33de